This change reorganizes the names of some classes related to the scheduler.
nano::scheduler::buckets has been renamed nano::scheduler::priority in order to match other election initating strategies like nano::scheduler::optimistic and nano::scheduler::hinted.

The class nano::prioritization has been renamed to nano::scheduler::buckets placing it under a more appropriate namespace and better indicating what the class is modeling.
