+++ nano/node/process_live_dispatcher.hpp
@@ -12,14 +12,14 @@ class transaction;
 
 namespace scheduler
 {
-	class buckets;
+	class priority;
 }
 
 // Observes confirmed blocks and dispatches the process_live function.
 class process_live_dispatcher
 {
 public:
-	process_live_dispatcher (nano::ledger & ledger, nano::scheduler::buckets & scheduler, nano::vote_cache & inactive_vote_cache, nano::websocket_server & websocket);
+	process_live_dispatcher (nano::ledger & ledger, nano::scheduler::priority & scheduler, nano::vote_cache & inactive_vote_cache, nano::websocket_server & websocket);
 	void connect (nano::block_processor & block_processor);
 
 private:
@@ -28,7 +28,7 @@ private:
 	void process_live (nano::block const & block, nano::transaction const & transaction);
 
 	nano::ledger & ledger;
-	nano::scheduler::buckets & scheduler;
+	nano::scheduler::priority & scheduler;
 	nano::vote_cache & inactive_vote_cache;
 	nano::websocket_server & websocket;
 };
