+++ nano/node/scheduler/buckets.cpp
@@ -1,172 +1,170 @@
-#include <nano/node/node.hpp>
+#include <nano/lib/blocks.hpp>
+#include <nano/lib/utility.hpp>
 #include <nano/node/scheduler/buckets.hpp>
 
-nano::scheduler::buckets::buckets (nano::node & node_a, nano::stats & stats_a) :
-	node{ node_a },
-	stats{ stats_a }
+#include <string>
+
+bool nano::scheduler::buckets::value_type::operator< (value_type const & other_a) const
 {
+	return time < other_a.time || (time == other_a.time && block->hash () < other_a.block->hash ());
 }
 
-nano::scheduler::buckets::~buckets ()
+bool nano::scheduler::buckets::value_type::operator== (value_type const & other_a) const
 {
-	// Thread must be stopped before destruction
-	debug_assert (!thread.joinable ());
+	return time == other_a.time && block->hash () == other_a.block->hash ();
 }
 
-void nano::scheduler::buckets::start ()
+/** Moves the bucket pointer to the next bucket */
+void nano::scheduler::buckets::next ()
 {
-	debug_assert (!thread.joinable ());
-
-	thread = std::thread{ [this] () {
-		nano::thread_role::set (nano::thread_role::name::election_scheduler);
-		run ();
-	} };
-}
-
-void nano::scheduler::buckets::stop ()
-{
+	++current;
+	if (current == schedule.end ())
 	{
-		nano::lock_guard<nano::mutex> lock{ mutex };
-		stopped = true;
+		current = schedule.begin ();
 	}
-	notify ();
-	nano::join_or_pass (thread);
 }
 
-void nano::scheduler::buckets::manual (std::shared_ptr<nano::block> const & block_a, boost::optional<nano::uint128_t> const & previous_balance_a, nano::election_behavior election_behavior_a)
+/** Seek to the next non-empty bucket, if one exists */
+void nano::scheduler::buckets::seek ()
 {
-	nano::lock_guard<nano::mutex> lock{ mutex };
-	manual_queue.push_back (std::make_tuple (block_a, previous_balance_a, election_behavior_a));
-	notify ();
-}
-
-bool nano::scheduler::buckets::activate (nano::account const & account_a, nano::transaction const & transaction)
-{
-	debug_assert (!account_a.is_zero ());
-	auto info = node.ledger.account_info (transaction, account_a);
-	if (info)
+	next ();
+	for (std::size_t i = 0, n = schedule.size (); buckets_m[*current].queue.empty () && i < n; ++i)
 	{
-		nano::confirmation_height_info conf_info;
-		node.store.confirmation_height.get (transaction, account_a, conf_info);
-		if (conf_info.height < info->block_count)
+		next ();
+	}
+}
+
+/** Initialise the schedule vector */
+void nano::scheduler::buckets::populate_schedule ()
+{
+	for (auto i = 0; i < buckets_m.size (); ++i)
+	{
+		schedule.push_back (i);
+	}
+}
+
+/**
+ * Prioritization constructor, construct a container containing approximately 'maximum' number of blocks.
+ * @param maximum number of blocks that this container can hold, this is a soft and approximate limit.
+ */
+nano::scheduler::buckets::buckets (uint64_t maximum) :
+	maximum{ maximum }
+{
+	auto build_region = [this] (uint128_t const & begin, uint128_t const & end, size_t count) {
+		auto width = (end - begin) / count;
+		for (auto i = 0; i < count; ++i)
 		{
-			debug_assert (conf_info.frontier != info->head);
-			auto hash = conf_info.height == 0 ? info->open_block : node.store.block.successor (transaction, conf_info.frontier);
-			auto block = node.store.block.get (transaction, hash);
-			debug_assert (block != nullptr);
-			if (node.ledger.dependents_confirmed (transaction, *block))
-			{
-				stats.inc (nano::stat::type::election_scheduler, nano::stat::detail::activated);
-				auto balance = node.ledger.balance (transaction, hash);
-				auto previous_balance = node.ledger.balance (transaction, conf_info.frontier);
-				nano::lock_guard<nano::mutex> lock{ mutex };
-				priority.push (info->modified, block, std::max (balance, previous_balance));
-				notify ();
-				return true; // Activated
-			}
+			minimums.push_back (begin + i * width);
 		}
+	};
+	minimums.push_back (uint128_t{ 0 });
+	build_region (uint128_t{ 1 } << 88, uint128_t{ 1 } << 92, 2);
+	build_region (uint128_t{ 1 } << 92, uint128_t{ 1 } << 96, 4);
+	build_region (uint128_t{ 1 } << 96, uint128_t{ 1 } << 100, 8);
+	build_region (uint128_t{ 1 } << 100, uint128_t{ 1 } << 104, 16);
+	build_region (uint128_t{ 1 } << 104, uint128_t{ 1 } << 108, 16);
+	build_region (uint128_t{ 1 } << 108, uint128_t{ 1 } << 112, 8);
+	build_region (uint128_t{ 1 } << 112, uint128_t{ 1 } << 116, 4);
+	build_region (uint128_t{ 1 } << 116, uint128_t{ 1 } << 120, 2);
+	minimums.push_back (uint128_t{ 1 } << 120);
+	buckets_m.resize (minimums.size ());
+	populate_schedule ();
+	current = schedule.begin ();
+}
+
+std::size_t nano::scheduler::buckets::index (nano::uint128_t const & balance) const
+{
+	auto index = std::upper_bound (minimums.begin (), minimums.end (), balance) - minimums.begin () - 1;
+	return index;
+}
+
+/**
+ * Push a block and its associated time into the prioritization container.
+ * The time is given here because sideband might not exist in the case of state blocks.
+ */
+void nano::scheduler::buckets::push (uint64_t time, std::shared_ptr<nano::block> block, nano::amount const & priority)
+{
+	auto was_empty = empty ();
+	auto & bucket = buckets_m[index (priority.number ())];
+	bucket.queue.emplace (value_type{ time, block });
+	if (bucket.queue.size () > std::max (decltype (maximum){ 1 }, maximum / buckets_m.size ()))
+	{
+		bucket.queue.erase (--bucket.queue.end ());
 	}
-	return false; // Not activated
+	if (was_empty)
+	{
+		seek ();
+	}
 }
 
-void nano::scheduler::buckets::flush ()
+/** Return the highest priority block of the current bucket */
+std::shared_ptr<nano::block> nano::scheduler::buckets::top () const
 {
-	nano::unique_lock<nano::mutex> lock{ mutex };
-	condition.wait (lock, [this] () {
-		return stopped || empty_locked () || node.active.vacancy () <= 0;
-	});
+	debug_assert (!empty ());
+	debug_assert (!buckets_m[*current].queue.empty ());
+	auto result = buckets_m[*current].queue.begin ()->block;
+	return result;
 }
 
-void nano::scheduler::buckets::notify ()
+/** Pop the current block from the container and seek to the next block, if it exists */
+void nano::scheduler::buckets::pop ()
 {
-	condition.notify_all ();
+	debug_assert (!empty ());
+	debug_assert (!buckets_m[*current].queue.empty ());
+	auto & bucket = buckets_m[*current];
+	bucket.queue.erase (bucket.queue.begin ());
+	seek ();
 }
 
+/** Returns the total number of blocks in buckets */
 std::size_t nano::scheduler::buckets::size () const
 {
-	nano::lock_guard<nano::mutex> lock{ mutex };
-	return priority.size () + manual_queue.size ();
+	std::size_t result{ 0 };
+	for (auto const & bucket : buckets_m)
+	{
+		result += bucket.queue.size ();
+	}
+	return result;
 }
 
-bool nano::scheduler::buckets::empty_locked () const
+/** Returns number of buckets, 62 by default */
+std::size_t nano::scheduler::buckets::bucket_count () const
 {
-	return priority.empty () && manual_queue.empty ();
+	return buckets_m.size ();
 }
 
+/** Returns number of items in bucket with index 'index' */
+std::size_t nano::scheduler::buckets::bucket_size (std::size_t index) const
+{
+	return buckets_m[index].queue.size ();
+}
+
+/** Returns true if all buckets are empty */
 bool nano::scheduler::buckets::empty () const
 {
-	nano::lock_guard<nano::mutex> lock{ mutex };
-	return empty_locked ();
+	return std::all_of (buckets_m.begin (), buckets_m.end (), [] (auto const & bucket_a) { return bucket_a.queue.empty (); });
 }
 
-std::size_t nano::scheduler::buckets::priority_queue_size () const
+/** Print the state of the class in stderr */
+void nano::scheduler::buckets::dump () const
 {
-	return priority.size ();
-}
-
-bool nano::scheduler::buckets::priority_queue_predicate () const
-{
-	return node.active.vacancy () > 0 && !priority.empty ();
-}
-
-bool nano::scheduler::buckets::manual_queue_predicate () const
-{
-	return !manual_queue.empty ();
-}
-
-void nano::scheduler::buckets::run ()
-{
-	nano::unique_lock<nano::mutex> lock{ mutex };
-	while (!stopped)
+	for (auto const & i : buckets_m)
 	{
-		condition.wait (lock, [this] () {
-			return stopped || priority_queue_predicate () || manual_queue_predicate ();
-		});
-		debug_assert ((std::this_thread::yield (), true)); // Introduce some random delay in debug builds
-		if (!stopped)
+		for (auto const & j : i.queue)
 		{
-			stats.inc (nano::stat::type::election_scheduler, nano::stat::detail::loop);
-
-			if (manual_queue_predicate ())
-			{
-				auto const [block, previous_balance, election_behavior] = manual_queue.front ();
-				manual_queue.pop_front ();
-				lock.unlock ();
-				stats.inc (nano::stat::type::election_scheduler, nano::stat::detail::insert_manual);
-				node.active.insert (block, election_behavior);
-			}
-			else if (priority_queue_predicate ())
-			{
-				auto block = priority.top ();
-				priority.pop ();
-				lock.unlock ();
-				stats.inc (nano::stat::type::election_scheduler, nano::stat::detail::insert_priority);
-				auto result = node.active.insert (block);
-				if (result.inserted)
-				{
-					stats.inc (nano::stat::type::election_scheduler, nano::stat::detail::insert_priority_success);
-				}
-				if (result.election != nullptr)
-				{
-					result.election->transition_active ();
-				}
-			}
-			else
-			{
-				lock.unlock ();
-			}
-			notify ();
-			lock.lock ();
+			std::cerr << j.time << ' ' << j.block->hash ().to_string () << '\n';
 		}
 	}
+	std::cerr << "current: " << std::to_string (*current) << '\n';
 }
 
 std::unique_ptr<nano::container_info_component> nano::scheduler::buckets::collect_container_info (std::string const & name)
 {
-	nano::unique_lock<nano::mutex> lock{ mutex };
-
 	auto composite = std::make_unique<container_info_composite> (name);
-	composite->add_component (std::make_unique<container_info_leaf> (container_info{ "manual_queue", manual_queue.size (), sizeof (decltype (manual_queue)::value_type) }));
-	composite->add_component (priority.collect_container_info ("priority"));
+	for (auto i = 0; i < buckets_m.size (); ++i)
+	{
+		auto const & bucket = buckets_m[i];
+		composite->add_component (std::make_unique<container_info_leaf> (container_info{ std::to_string (i), bucket.queue.size (), 0 }));
+	}
 	return composite;
 }
