+++ nano/node/scheduler/component.hpp
@@ -2,23 +2,26 @@
 
 #include <memory>
 
+namespace nano
+{
+class node;
+}
 namespace nano::scheduler
 {
-class buckets;
+class priority;
 class hinted;
-class node;
 class optimistic;
 
 class component
 {
 	std::unique_ptr<nano::scheduler::optimistic> optimistic_impl;
-	std::unique_ptr<nano::scheduler::buckets> buckets_impl;
+	std::unique_ptr<nano::scheduler::priority> priority_impl;
 	std::unique_ptr<nano::scheduler::hinted> hinted_impl;
 
 public:
 	explicit component (nano::node & node);
 
-	nano::scheduler::buckets & buckets;
+	nano::scheduler::priority & priority;
 	nano::scheduler::hinted & hinted;
 	nano::scheduler::optimistic & optimistic;
 };
